Prepare for Swarm support

+++ include/checksum.hpp
@@ -3,4 +3,4 @@
 #include <Arduino.h>
 #include <stdint.h>
 
-uint8_t getChecksum(double arr[], uint16_t length);
+uint8_t getChecksum(int32_t arr[], uint16_t length);
